--- a/src/dense/high_order_rk_addsteps.jl
+++ b/src/dense/high_order_rk_addsteps.jl
@@ -40,7 +40,6 @@ end
     @unpack d401,d406,d407,d408,d409,d410,d411,d412,d413,d414,d415,d416,d501,d506,d507,d508,d509,d510,d511,d512,d513,d514,d515,d516,d601,d606,d607,d608,d609,d610,d611,d612,d613,d614,d615,d616,d701,d706,d707,d708,d709,d710,d711,d712,d713,d714,d715,d716 = cache.tab
     @unpack k1,k2,k3,k4,k5,k6,k7,k8,k9,k10,k11,k12,k13,k14,k15,k16,udiff,bspl,dense_tmp3,dense_tmp4,dense_tmp5,dense_tmp6,dense_tmp7,kupdate,utilde,tmp = cache
     utmp = utilde
-    k = [cache.udiff,cache.bspl,cache.dense_tmp3,cache.dense_tmp4,cache.dense_tmp5,cache.dense_tmp6,cache.dense_tmp7]
     f(k1,uprev,p,t)
     @.. tmp = uprev+dt*(a0201*k1)
     f(k2,tmp,p,t+c2*dt)
@@ -74,12 +73,19 @@ end
     @.. tmp = uprev+dt*(a1601*k1+a1606*k6+a1607*k7+a1608*k8+a1609*k9+a1613*k13+a1614*k14+a1615*k15)
     f(k16,tmp,p,t+c16*dt)
     copyto!(udiff, kupdate)
-    @.. bspl = k1 - udiff
-    @.. k[3] = udiff - k13 - bspl
-    @.. k[4] = (d401*k1+d406*k6+d407*k7+d408*k8+d409*k9+d410*k10+d411*k11+d412*k12+d413*k13+d414*k14+d415*k15+d416*k16)
-    @.. k[5] = (d501*k1+d506*k6+d507*k7+d508*k8+d509*k9+d510*k10+d511*k11+d512*k12+d513*k13+d514*k14+d515*k15+d516*k16)
-    @.. k[6] = (d601*k1+d606*k6+d607*k7+d608*k8+d609*k9+d610*k10+d611*k11+d612*k12+d613*k13+d614*k14+d615*k15+d616*k16)
-    @.. k[7] = (d701*k1+d706*k6+d707*k7+d708*k8+d709*k9+d710*k10+d711*k11+d712*k12+d713*k13+d714*k14+d715*k15+d716*k16)
+    copyat_or_push!(k,1,udiff)
+    @.. tmp = k1 - udiff
+    copyat_or_push!(k,2,tmp)
+    @.. tmp = udiff - k13 - bspl
+    copyat_or_push!(k,3,tmp)
+    @.. tmp = (d401*k1+d406*k6+d407*k7+d408*k8+d409*k9+d410*k10+d411*k11+d412*k12+d413*k13+d414*k14+d415*k15+d416*k16)
+    copyat_or_push!(k,4,tmp)
+    @.. tmp = (d501*k1+d506*k6+d507*k7+d508*k8+d509*k9+d510*k10+d511*k11+d512*k12+d513*k13+d514*k14+d515*k15+d516*k16)
+    copyat_or_push!(k,5,tmp)
+    @.. tmp = (d601*k1+d606*k6+d607*k7+d608*k8+d609*k9+d610*k10+d611*k11+d612*k12+d613*k13+d614*k14+d615*k15+d616*k16)
+    copyat_or_push!(k,6,tmp)
+    @.. tmp = (d701*k1+d706*k6+d707*k7+d708*k8+d709*k9+d710*k10+d711*k11+d712*k12+d713*k13+d714*k14+d715*k15+d716*k16)
+    copyat_or_push!(k,7,tmp)
   end
 end