--- a/src/dense/rosenbrock_interpolants.jl
+++ b/src/dense/rosenbrock_interpolants.jl
@@ -15,9 +15,14 @@ end
   c2 = Θ*(Θ-2d)/(1-2d)
 end
 
-@muladd function _ode_interpolant(Θ,dt,y₀,y₁,k,cache::Union{Rosenbrock23ConstantCache,Rosenbrock23Cache,Rosenbrock32ConstantCache,Rosenbrock32Cache},idxs::Nothing,T::Type{Val{0}})
+@muladd function _ode_interpolant(Θ,dt,y₀,y₁,k,cache::Union{Rosenbrock23ConstantCache,Rosenbrock32ConstantCache},idxs::Nothing,T::Type{Val{0}})
   @rosenbrock2332pre0
-  @.. y₀ + dt*(c1*k[1] + c2*k[2])
+  @inbounds y₀ + dt*(c1*k[1] + c2*k[2])
+end
+
+@muladd function _ode_interpolant(Θ,dt,y₀,y₁,k,cache::Union{Rosenbrock23Cache,Rosenbrock32Cache},idxs::Nothing,T::Type{Val{0}})
+  @rosenbrock2332pre0
+  @inbounds @.. y₀ + dt*(c1*k[1] + c2*k[2])
 end
 
 @muladd function _ode_interpolant(Θ,dt,y₀,y₁,k,cache::Union{Rosenbrock23ConstantCache,Rosenbrock23Cache,Rosenbrock32ConstantCache,Rosenbrock32Cache},idxs,T::Type{Val{0}})
@@ -27,7 +32,7 @@ end
 
 @muladd function _ode_interpolant!(out,Θ,dt,y₀,y₁,k,cache::Union{Rosenbrock23ConstantCache,Rosenbrock23Cache,Rosenbrock32ConstantCache,Rosenbrock32Cache},idxs::Nothing,T::Type{Val{0}})
   @rosenbrock2332pre0
-  @.. out = y₀ + dt*(c1*k[1] + c2*k[2])
+  @inbounds @.. out = y₀ + dt*(c1*k[1] + c2*k[2])
   out
 end
 
@@ -69,9 +74,14 @@ end
 """
 From MATLAB ODE Suite by Shampine
 """
-@muladd function _ode_interpolant(Θ,dt,y₀,y₁,k,cache::Union{Rodas4ConstantCache,Rodas4Cache},idxs::Nothing,T::Type{Val{0}})
+@muladd function _ode_interpolant(Θ,dt,y₀,y₁,k,cache::Rodas4ConstantCache,idxs::Nothing,T::Type{Val{0}})
   Θ1 = 1 - Θ
-  @.. Θ1*y₀ + Θ*(y₁ + Θ1*(k[1] + Θ*k[2]))
+  @inbounds Θ1*y₀ + Θ*(y₁ + Θ1*(k[1] + Θ*k[2]))
+end
+
+@muladd function _ode_interpolant(Θ,dt,y₀,y₁,k,cache::Rodas4Cache,idxs::Nothing,T::Type{Val{0}})
+  Θ1 = 1 - Θ
+  @inbounds @.. Θ1*y₀ + Θ*(y₁ + Θ1*(k[1] + Θ*k[2]))
 end
 
 @muladd function _ode_interpolant(Θ,dt,y₀,y₁,k,cache::Union{Rodas4ConstantCache,Rodas4Cache},idxs,T::Type{Val{0}})