This PR builds on https://github.com/JuliaDiffEq/OrdinaryDiffEq.jl/pull/988, but changes the computation of the interpolation intervals. It covers solutions with only one time point and fixes https://github.com/JuliaDiffEq/OrdinaryDiffEq.jl/issues/991.