@ChrisRackauckas 2020-01-11
```julia
using OrdinaryDiffEq, StaticArrays, BenchmarkTools
function lorenz_static(u,p,t)
    @inbounds begin
        dx = p[1]*(u[2]-u[1])
        dy = u[1]*(p[2]-u[3]) - u[2]
        dz = u[1]*u[2] - p[3]*u[3]
    end
    @SVector [dx,dy,dz]
end
u0 = @SVector [1.0,0.0,0.0]
p  = @SVector [10.0,28.0,8/3]
tspan = (0.0,100.0)
prob = ODEProblem(lorenz_static,u0,tspan,p)
sol = solve(prob,Tsit5(),saveat=0.1)
@btime sol(1.55)
@btime sol(1.5)

# Master:
1.420 μs (38 allocations: 1.14 KiB)
974.359 ns (28 allocations: 976 bytes)

# PR:
565.574 ns (20 allocations: 864 bytes)
565.574 ns (20 allocations: 864 bytes)

function lorenz(du,u,p,t)
 du[1] = 10.0(u[2]-u[1])
 du[2] = u[1]*(28.0-u[3]) - u[2]
 du[3] = u[1]*u[2] - (8/3)*u[3]
end
u0 = [1.0;0.0;0.0]
tspan = (0.0,100.0)
prob = ODEProblem(lorenz,u0,tspan)
sol = solve(prob,Tsit5(),saveat=0.1)
@btime sol(1.55)
@btime sol(1.5)

# Master:
1.360 μs (33 allocations: 624 bytes)
872.881 ns (24 allocations: 384 bytes)

# PR:
571.429 ns (17 allocations: 368 bytes)
713.963 ns (17 allocations: 368 bytes)
```

The fix makes it no longer fix https://github.com/JuliaDiffEq/OrdinaryDiffEq.jl/issues/991